# src/api-doc.yml
openapi: 3.0.3
info:
  title: 'ContreauL API'
  version: '1.0.0'
tags:
  - name: Users
    description: Manage users access
  - name: Weather
    description: Provide weather forecast information's
  - name: Logs
    description: Provide logs management
paths:
  /v1/users:
    post:
      tags:
        - Users
      summary: Add a new user
      requestBody:
        description: The user to add
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - password
                - email
              properties:
                name:
                  type: string
                  description: The user name
                password:
                  type: string
                  description: The user password
                email:
                  type: string
                  description: The user email
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: integer
                    description: The user Token
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSHVndWVzIiwiZW1haWwiOiJodWdlczYyQGdtYWlsLmNvbSJ9.hjIs6IVN1WSh5IuyoyIt85zBKFk_cmnRPsDvUlcHQ6U
        '401':
          description: User already exist
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message
                    example: User already exist
  /v1/users/login:
    post:
      tags:
        - Users
      summary: Login a user
      requestBody:
        description: The user to login
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - email
              properties:
                password:
                  type: string
                  description: The user password
                email:
                  type: string
                  description: The user email
      responses:
        '200':
          description: Logged in successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: integer
                    description: The user Token
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJuYW1lIjoiSHVndWVzIiwiZW1haWwiOiJodWdlczYyQGdtYWlsLmNvbSJ9.hjIs6IVN1WSh5IuyoyIt85zBKFk_cmnRPsDvUlcHQ6U
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message
                    example: Invalid email or password
  /v1/weather/today:
    get:
      tags:
        - Weather
      summary: Provide weather forecast for today
      parameters:
        - in: query
          name: insee
          schema:
            type: integer
          required: true
          description: City insee number
          example: 62500
      responses:
        '200':
          description: Weather forecast retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  weather:
                    type: array
                    minItems: 4
                    maxItems: 4
                    items:
                      type: object
                      properties:
                        forecast:
                          type: object
        '400':
          description: Insee is missing in query
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message
                    example: Insee query param is missing !
  /v1/weather/today/rain:
    get:
      tags:
        - Weather
      summary: Provide raining boolean for today
      parameters:
        - in: query
          name: insee
          schema:
            type: integer
          required: true
          description: City insee number
          example: 62500
      responses:
        '200':
          description: Weather forecast retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  raining:
                    type: boolean
        '400':
          description: Insee is missing in query
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message
                    example: Insee query param is missing !
  /v1/logs:
    post:
      tags:
        - Logs
      summary: Add a new log
      requestBody:
        description: The log to add
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - device
                - water_temperature
                - water_level
                - occurred_at
                - ph
              properties:
                device:
                  type: integer
                  description: Device associated to the log
                water_temperature:
                  type: integer
                  description: The water temperature value
                water_level:
                  type: string
                  description: The water level value
                occurred_at:
                  type: string
                  description: Date when event happened
                ph:
                  type: integer
                  description: The ph value
      responses:
        '201':
          description: Log created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: The log id
                    example: 71f2770b-bd5f-413a-ab74-8d6fcf26f6db
        '401':
          description: Invalid given data
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: The error message
                    example: Invalid data !